requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3
py3exiv2==0.12.0
pillow==11.0.0
matplotlib==3.9.4
opencv-python==4.10.0.84